requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF>=1.2.0
Flask-Caching==2.1.0
celery==5.3.6
redis==5.0.1
openai>=1.0.0
gunicorn==21.2.0
email-validator==2.0.0
python-dotenv==1.0.0